Add files via upload
I've been given a problem statement:

A university is receiving applications from students to register for language courses offered in six different languages. Each language has limited seats and time of submission of request is to be used for making a decision. The following are the fields provided in the application:
a.	Register Number, Name, Time and date of submission, Preference of language {Option1, Option2, Option3}
b.	Devise a solution that would automate the process of allocating the languages based on the order of preference given by students. Use appropriate data structures to process the data. Include provision for swapping of the language that has been allocated for a student on mutual consent. 

I have prepared a code using Gale Shapely Algorithm to select the best preference in C.
However, the problem says that the user's time and date of entering the preference will be recorded which will be used in decision making. I don't know which data structure to implement and how to ask for input from the user and then convert it in a 2D array to process gale shapely algorithm.